03 - Data Science and Custom Skills/Docker Flask Skill/requirements.txt
Flask==1.0.2
pyopenssl==19.0.0
click==6.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
jsonify==0.5
azure-storage==0.36.0
nltk==3.4.1
spacy==2.1.4
pandas==0.24.2
